+++ src/Ampersand/ADL1/Expression.hs
@@ -4,6 +4,7 @@ module Ampersand.ADL1.Expression (
                       ,subst
                       ,primitives, subExpressions, isMp1, isEEps, isEDcD
                       ,isPos,isNeg, deMorganERad, deMorganECps, deMorganEUni, deMorganEIsc, notCpl, isCpl, isFlipped
+                      ,isFitForCrudC ,isFitForCrudR ,isFitForCrudU ,isFitForCrudD
                       ,exprIsc2list, exprUni2list, exprCps2list, exprRad2list, exprPrd2list
                       ,insParentheses)
 where
@@ -146,6 +147,72 @@ isFlipped :: Expression -> Bool
 isFlipped EFlp{}   = True
 isFlipped (EBrk e) = isFlipped e
 isFlipped _        = False
+
+-- | Function to determine that the expression
+--   could be used to create a new atom in its target concept
+isFitForCrudC :: Expression -> Bool
+isFitForCrudC expr = 
+   case expr of 
+     EDcD{}   -> True
+     EFlp e   -> isFitForCrudC e
+     EBrk e   -> isFitForCrudC e
+     EEps _ _ -> False
+     EDcI{}   -> True -- TODO: set to False when functionality of +menu is adapted from I[Cpt] to V[SESSION*Cpt] expressions (see Issue #884)
+     EMp1{}   -> False
+     EDcV{}   -> True
+     ECps ( (EEps _ _), e         ) -> isFitForCrudC e
+     ECps ( e         , (EEps _ _)) -> isFitForCrudC e
+     ECps ( _         , _         ) -> True
+     EEqu{}   -> True
+     EInc{}   -> True
+     EIsc{}   -> True
+     EUni{}   -> True
+     EDif{}   -> True
+     ELrs{}   -> True
+     ERrs{}   -> True
+     EDia{}   -> True
+     ERad{}   -> True
+     EPrd{}   -> True
+     EKl0{}   -> True
+     EKl1{}   -> True
+     ECpl{}   -> True
+-- | Function to determine that the expression
+--   could be used to read the population of its target concept
+isFitForCrudR :: Expression -> Bool
+isFitForCrudR _ = True 
+-- | Function to determine that the expression
+--   could be used to insert or delete a pair in the population of a relation
+isFitForCrudU :: Expression -> Bool
+isFitForCrudU expr = 
+   case expr of 
+     EDcD{}   -> True
+     EFlp e   -> isFitForCrudU e
+     EBrk e   -> isFitForCrudU e
+     EEps _ _ -> False
+     EDcI{}   -> False
+     EMp1{}   -> False
+     EDcV{}   -> False
+     ECps ( (EEps _ _), e         ) -> isFitForCrudU e
+     ECps ( e         , (EEps _ _)) -> isFitForCrudU e
+     ECps ( e         , EDcI{}    ) -> isFitForCrudU e
+     ECps ( _         , _         ) -> False
+     EEqu{}   -> False
+     EInc{}   -> False
+     EIsc{}   -> False
+     EUni{}   -> False
+     EDif{}   -> False
+     ELrs{}   -> False
+     ERrs{}   -> False
+     EDia{}   -> False
+     ERad{}   -> False
+     EPrd{}   -> False
+     EKl0{}   -> False
+     EKl1{}   -> False
+     ECpl{}   -> False
+-- | Function to determine that the expression is simple, that it
+--   could be used to update the population of a relation
+isFitForCrudD :: Expression -> Bool
+isFitForCrudD _ = True
 
 
 exprIsc2list, exprUni2list, exprCps2list, exprRad2list, exprPrd2list :: Expression -> [Expression]