+++ src/Ampersand/Test/RunAmpersand.hs
@@ -19,4 +19,4 @@ runAmpersand opts file =
            case gFSpec of
               Errors err    -> return $ NEL.toList err
               --TODO: Do something with the fSpec
-              Checked _     -> return []
+              Checked _ _   -> return []